+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ### [Unreleased][unreleased]
 
 #### Added
+- json output for commands
 - migration datetime to Status command
-- json output for Status command
 
 ### [0.7.0] - 2017-04-06
 #### Changed